+++ src/scenario/custom_messages.c
@@ -315,7 +315,7 @@ const uint8_t *custom_messages_get_background_image(custom_message_t *message)
     }
 }
 
-int custom_messages_relink_text_blob(int text_id, text_blob_string_t *new_text_link)
+int custom_messages_relink_text_blob(size_t text_id, text_blob_string_t *new_text_link)
 {
     custom_message_t *entry;
     array_foreach(custom_messages, entry) {
